refactor(engine): Rename utils to identifiers
